+++ Zero/Sources/Network/Models/Request/ZWalletTransferToken.swift
@@ -11,9 +11,9 @@ public struct ZWalletTransferToken: Encodable {
     public let to: String
     public let amount: String
     public let tokenAddress: String
-    public let chainId: Int
+    public let chainId: UInt64
     
-    init(recipientWalletAddress: String, amount: String, tokenAddress: String, chainId: Int) {
+    init(recipientWalletAddress: String, amount: String, tokenAddress: String, chainId: UInt64) {
         self.to = recipientWalletAddress
         self.amount = amount
         self.tokenAddress = tokenAddress
